requirements.txt
request==2019.4.13
requests==2.20.0
Flask==1.0.2
Pyrebase==3.0.27 
regex==2019.08.19
gunicorn==19.9.0
psycopg2==2.8.4
PyJWT==1.7.1